tests: Refactor to real tempdir
For some upcoming tests it's important that they temp dir is not in a
SCM. So placing the tempdir in 'tests/TestCmdAddginlx7h6/' does not work
anymore, because this is inside the git repo of "supatch" itself.

So add new helper that uses "TMPDIR" or something, e.g. under "/tmp/"